+++ packages/fuselage/src/components/Box/index.d.ts
@@ -191,7 +191,6 @@ export const Box: {
 
 export { default as AnimatedVisibility } from './AnimatedVisibility';
 export { default as Position, PositionAnimated } from './Position';
-export { default as Scrollable } from './Scrollable';
 
 export const useArrayLikeClassNameProp: <
   T extends {
