+++ packages/fuselage/src/components/Box/index.js
@@ -106,4 +106,3 @@ Box.displayName = 'Box';
 
 export { default as AnimatedVisibility } from './AnimatedVisibility';
 export { default as Position, PositionAnimated } from './Position';
-export { default as Scrollable } from './Scrollable';
