+++ packages/fuselage/src/components/Scrollable/Scrollable.spec.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 
-import Scrollable from '.';
+import { Scrollable } from '../..';
 
 it('renders without crashing', () => {
   render(<Scrollable />);
